move debug class to debug.py
