+++ pman.py
@@ -46,123 +46,20 @@ import  socket
 import  queue
 from    functools       import  partial
 import  inspect
-import  logging
+import  platform
 
+import  multiprocessing
+
+import  pudb
+
+# pman local dependencies
 sys.path.append(os.path.join(os.path.dirname(__file__), './'))
 import  crunner
 import  C_snode
 import  message
 from    _colors         import  Colors
 import  pfioh
-
-import  platform
-import  multiprocessing
-
-import  pudb
-
-
-logging.basicConfig(level=logging.DEBUG,
-                    format='(%(threadName)-10s) %(message)s')
-
-class debug(object):
-    """
-        A simple class that provides some helper debug functions. Mostly
-        printing function/thread names and checking verbosity level
-        before printing.
-    """
-
-    def log(self, *args):
-        """
-        get/set the log object.
-
-        Caller can further manipulate the log object with object-specific
-        calls.
-        """
-        if len(args):
-            self._log = args[0]
-        else:
-            return self._log
-
-    def name(self, *args):
-        """
-        get/set the descriptive name text of this object.
-        """
-        if len(args):
-            self.__name = args[0]
-        else:
-            return self.__name
-
-    def __init__(self, **kwargs):
-        """
-        Constructor
-        """
-
-        self.verbosity  = 0
-        self.level      = 0
-
-        self.b_useDebug             = False
-        self.str_debugDirFile       = '/tmp'
-        for k, v in kwargs.items():
-            if k == 'verbosity':    self.verbosity          = v
-            if k == 'level':        self.level              = v
-            if k == 'debugToFile':  self.b_useDebug         = v
-            if k == 'debugFile':    self.str_debugDirFile   = v
-
-        if self.b_useDebug:
-            str_debugDir                = os.path.dirname(self.str_debugDirFile)
-            str_debugName               = os.path.basename(self.str_debugDirFile)
-            if not os.path.exists(str_debugDir):
-                os.makedirs(str_debugDir)
-            self.str_debugFile          = '%s/%s' % (str_debugDir, str_debugName)
-            self.debug                  = message.Message(logTo = self.str_debugFile)
-            self.debug._b_syslog        = False
-            self.debug._b_flushNewLine  = True
-        self._log                   = message.Message()
-        self._log._b_syslog         = True
-        self.__name                 = "pman"
-
-
-    def __call__(self, *args, **kwargs):
-        self.qprint(*args, **kwargs)
-
-    def qprint(self, *args, **kwargs):
-        """
-        The "print" command for this object.
-
-        :param kwargs:
-        :return:
-        """
-
-        self.level  = 0
-        self.msg    = ""
-
-        for k, v in kwargs.items():
-            if k == 'level':    self.level  = v
-            if k == 'msg':      self.msg    = v
-
-        if len(args):
-            self.msg    = args[0]
-
-        if self.b_useDebug:
-            write   = self.debug
-        else:
-            write   = print
-
-        if self.level <= self.verbosity:
-
-            if self.b_useDebug:
-                write('| %50s | %30s | ' % (
-                    threading.current_thread(),
-                    inspect.stack()[1][3]
-                ), end='', syslog = True)
-            else:
-                write('%26s | %50s | %30s | ' % (
-                    datetime.datetime.now(),
-                    threading.current_thread(),
-                    inspect.stack()[1][3]
-                ), end='')
-            for t in range(0, self.level): write("\t", end='')
-            write(self.msg)
+import  debug
 
 class StoreHandler(BaseHTTPRequestHandler):
 
@@ -304,7 +201,7 @@ class ThreadedHTTPServer(ThreadingMixIn, HTTPServer):
         self.b_removeZip    = False
         self.args           = None
 
-        self.dp             = debug(verbosity=0, level=-1)
+        self.dp             = debug.debug(verbosity=0, level=-1)
 
         for k,v in kwargs.items():
             if k == 'args': self.args   = v
@@ -387,7 +284,7 @@ class pman(object):
         # Screen formatting
         self.LC                 = 30
         self.RC                 = 50
-        self.dp                 = debug(    verbosity   = 0,
+        self.dp                 = debug.debug(    verbosity   = 0,
                                             level       = -1,
                                             debugFile   = self.str_debugFile,
                                             debugToFile = self.b_debugToFile)
@@ -650,7 +547,7 @@ class FileIO(threading.Thread):
     def __init__(self, **kwargs):
         self.__name             = "FileIO"
         self.b_http             = False
-        self.dp                 = debug(verbosity=0, level=-1)
+        self.dp                 = debug.debug(verbosity=0, level=-1)
 
         self.str_DBpath         = "/tmp/pman"
 
@@ -686,7 +583,7 @@ class Listener(threading.Thread):
     def __init__(self, **kwargs):
         self.__name             = "Listener"
         self.b_http             = False
-        self.dp                 = debug(verbosity=0, level=-1)
+        self.dp                 = debug.debug(verbosity=0, level=-1)
 
         self.poller             = None
         self.str_DBpath         = "/tmp/pman"
@@ -1529,7 +1426,7 @@ class Poller(threading.Thread):
 
         self.pollTime           = 10
 
-        self.dp                 = debug(verbosity=0, level=-1)
+        self.dp                 = debug.debug(verbosity=0, level=-1)
 
         self.str_cmd            = ""
         self.crunner            = None
@@ -1580,7 +1477,7 @@ class Crunner(threading.Thread):
 
     def __init__(self, **kwargs):
         self.__name             = "Crunner"
-        self.dp                 = debug(verbosity=0, level=-1)
+        self.dp                 = debug.debug(verbosity=0, level=-1)
 
         self.dp.qprint('starting crunner...', level=-1)
