+++ pman_client.py
@@ -28,6 +28,8 @@ import  pfioh
 import  urllib
 import  datetime
 import  codecs
+
+# pman local dependencies
 import  purl
 
 class Client():
